Create README.md
# Java_String_Student_Grade
This is Grading System Problem. we can check many of students average mark,passingRate,highest mark. 
You are tasked with creating a program to track the grades of students in a class.
Implemen ta Java program that does the following:
●Definean ArrayList to store the grades of students(integer values). 
●Write a method called addGrade that takes an integer(student's grade) as input and adds it to the ArrayList.
● Implement a method named calculateAverage that calculates and returns the average grade of all the students.
●Create a metho dcalled highest Gradethat finds and returns the highest grade in the ArrayList.
●Design a function named list PassingStudents that prints the grades of students who have passed(grades>=50).
●Write a main method to demonstrate the function a lity of your program.Prompt theuser to input grades until-1 is enter by the user,
then display the average grade,highest grade,and list of passing students.